+++ tunet-flutter/lib/views/about_page.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class AboutPage extends StatelessWidget {
+  const AboutPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+        margin: const EdgeInsets.all(8.0),
+        child: Column(
+          children: [
+            const FlutterLogo(),
+          ],
+        ));
+  }
+}
